requirements.txt
pandas
requests
psycopg2-binary
python-dotenv
json
logging